+++ Sources/TestHelper/NavNative/NativeNavigatorSpy.swift
@@ -33,7 +33,8 @@ public class NativeNavigatorSpy: MapboxNavigationNative.Navigator {
             utilizeSensorData: true,
             historyDirectoryURL: nil,
             initialManeuverAvoidanceRadius: 8,
-            locale: .current
+            locale: .current,
+            rerouteStrategyForMatchRoute: .rerouteDisabled
         )
 
         super.init(