+++ Sources/_MapboxNavigationTestHelpers/Core/RoutesDataMock.swift
@@ -1,0 +1,13 @@
+@preconcurrency import MapboxNavigationNative
+
+public class RoutesDataMock: RoutesData {
+    public init() {}
+
+    public func primaryRoute() -> any RouteInterface {
+        RouteInterfaceMock()
+    }
+
+    public func alternativeRoutes() -> [RouteAlternative] {
+        []
+    }
+}